Squashed 'osk-apps/kit_sch/' changes from efe5242..d7185a9
d7185a9 Updated tables

git-subtree-dir: osk-apps/kit_sch
git-subtree-split: d7185a9ff4d9c8e6d744c53a228cc5097622d6c0

+++ fsw/src/scheduler.c
@@ -637,8 +637,8 @@ bool SCHEDULER_SendMsgEntryCmd(void* ObjDataPtr, const CFE_MSG_Message_t *MsgPtr
          CFE_MSG_ValidateChecksum(MsgPtr, &ValidChecksum);
          
          CFE_EVS_SendEvent(SCHEDULER_CMD_SUCCESS_EID, CFE_EVS_EventType_INFORMATION, 
-                           "Msg[%d]=Command(ApId,SeqCnt,Len,FuncCode,ValidChecksum)=>(0x%04X,%d,%ld,%d,0x%02X)",
-                           MsgIndex,ApId,SeqCnt,Size,FuncCode,ValidChecksum);
+                           "Msg[%d]=Command(ApId,SeqCnt,Len,FuncCode,ValidChecksum)=>(0x%04X,%d,%d,%d,0x%02X)",
+                           MsgIndex,ApId,SeqCnt,(unsigned int)Size,FuncCode,ValidChecksum);
             
          DataBuf = &(Scheduler->MsgTbl.Data.Entry[MsgIndex].Buffer[sizeof(CFE_MSG_CommandHeader_t)/2]);
 
@@ -651,8 +651,8 @@ bool SCHEDULER_SendMsgEntryCmd(void* ObjDataPtr, const CFE_MSG_Message_t *MsgPtr
          CFE_MSG_GetMsgTime(MsgPtr, &Time);
          
          CFE_EVS_SendEvent(SCHEDULER_CMD_SUCCESS_EID, CFE_EVS_EventType_INFORMATION, 
-                           "Msg[%d]=Telemetry(ApId,SeqCnt,Len,Seconds,Subsecs)=>(0x%04X,%d,%ld,%d,%d)",
-                           MsgIndex,ApId,SeqCnt,Size,Time.Seconds,Time.Subseconds);
+                           "Msg[%d]=Telemetry(ApId,SeqCnt,Len,Seconds,Subsecs)=>(0x%04X,%d,%d,%d,%d)",
+                           MsgIndex,ApId,SeqCnt,(unsigned int)Size,Time.Seconds,Time.Subseconds);
            
          DataBuf = &(Scheduler->MsgTbl.Data.Entry[MsgIndex].Buffer[sizeof(CFE_MSG_TelemetryHeader_t)/2]);
